+++ Frends.AmazonS3.DownloadObject/Frends.AmazonS3.DownloadObject/Definitions/SingleResultObject.cs
@@ -6,27 +6,41 @@
 public class SingleResultObject
 {
     /// <summary>
-    /// Single download result.
-    /// </summary>
-    /// <example>Download complete: {fullPath}.</example>
-    public string ObjectData { get; private set; }
-
-    /// <summary>
     /// Object's name.
     /// </summary>
     /// <example>Filename.txt</example>
     public string ObjectName { get; private set; }
 
     /// <summary>
-    /// Full path.
+    /// Full destination path.
     /// </summary>
     /// <example>c:\temp\Filename.txt, \\network\folder\Filename.txt</example>
     public string FullPath { get; private set; }
 
-    internal SingleResultObject(string objectData, string objectName, string fullPath)
+    /// <summary>
+    /// Original destination file was overwritten.
+    /// </summary>
+    /// <example>false</example>
+    public bool Overwritten { get; private set; }
+
+    /// <summary>
+    /// Source object was deleted.
+    /// </summary>
+    /// <example>false</example>
+    public bool SourceDeleted { get; private set; }
+
+    /// <summary>
+    /// Additional information of this object.
+    /// </summary>
+    /// <example>Object skipped because file already exists in destination.</example>
+    public string Info { get; private set; }
+
+    internal SingleResultObject(string objectName, string fullPath, bool overwritten, bool sourceDeleted, string info)
     {
-        ObjectData = objectData;
         ObjectName = objectName;
         FullPath = fullPath;
+        Overwritten = overwritten;
+        SourceDeleted = sourceDeleted;
+        Info = info;
     }
 }