+++ Frends.AmazonS3.DownloadObject/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.0.0] - 2022-12-02
+### Modified
+- Memory leak fix.
+- SingleResultObject parameter changes:
+	- New parameters: Overwritten, SourceDeleted, Info.
+	- Removed parameter: ObjectData (replaced by Info). 
+
 ## [1.0.0] - 2022-05-09
 ### Added
 - Initial implementation